+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Preserve parentheses around infix function left-hand sides with extra arguments ([#1243])
+- Recognize and normalize CPP directives with spaces after `#` ([#1249])
 
 ### Removed
 
@@ -430,6 +431,7 @@ This version is accidentally pushlished, and is the same as 5.3.3.
 [@uhbif19]: https://github.com/uhbif19
 [@toku-sa-n]: https://github.com/toku-sa-n
 
+[#1249]: https://github.com/mihaimaruseac/hindent/pull/1249
 [#1243]: https://github.com/mihaimaruseac/hindent/pull/1243
 [#1163]: https://github.com/mihaimaruseac/hindent/pull/1163
 [#1143]: https://github.com/mihaimaruseac/hindent/pull/1143